+++ modules/custom_logger.py
@@ -1,0 +1,7 @@
+"""Custom logger"""
+import sys
+
+from loguru import logger
+
+logger.remove()
+logger.add(sys.stderr, format="<green>{time:YYYY-MM-DD HH:mm:ss}</green> | <lvl>{level}</lvl> | <lvl>{message}</lvl>")
